requirements.txt
fastapi==0.136.1
uvicorn[standard]==0.34.0
motor==3.7.0
pymongo==4.12.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
httpx==0.28.1
langchain-core==1.3.2
langgraph==1.1.10
langsmith==0.4.1
tiktoken==0.9.0
google-auth==2.40.1
google-auth-oauthlib==1.2.1
requests==2.32.3
pydantic-settings==2.9.1
python-dotenv==1.1.0
aiohttp==3.12.6
sse-starlette==2.3.4
pydantic[email]
langchain_ollama